authorise pour le territoire, l'équipe et l'agent

+++ app/controllers/admin/territories/agent_territorial_roles_controller.rb
@@ -13,12 +13,12 @@ class Admin::Territories::AgentTerritorialRolesController < Admin::Territories::
       .includes(:territories)
       .to_a
       .reject { _1.territorial_admin_in?(current_territory) }
-    authorize_admin(@role)
+    authorize @role
   end
 
   def create
     @role = AgentTerritorialRole.new(agent_territorial_role_params)
-    authorize_admin(@role)
+    authorize @role
     if @role.save
       redirect_to(
         admin_territory_agent_territorial_roles_path(current_territory),
@@ -31,7 +31,7 @@ class Admin::Territories::AgentTerritorialRolesController < Admin::Territories::
 
   def destroy
     role = AgentTerritorialRole.find(params[:id])
-    authorize_admin(role)
+    authorize role
     if role.destroy
       flash[:success] = "#{role.agent.full_name} n'a plus le rôle d'administrateur du #{current_territory}"
     else
